inference_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.21.4
transformers==4.35.0
accelerate==0.24.1
controlnet-aux==0.4.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
